Refactor duel logic into separate class
